Simplify actions for Breadcrumb component
