+++ sensemap/src/types/map-dispatch.d.ts
@@ -1,4 +1,25 @@
-import { Dispatch } from '.';
+import { Dispatch as ReduxDispatch, combineReducers } from 'redux';
+import * as SM from './sense-map';
+import * as SO from './sense-object';
+import * as SL from './selection';
+import * as OE from './object-editor';
+import * as V  from './viewport';
+import * as SG from './stage';
+import * as I  from './input';
+import * as IP from './importer';
+
+export type State = {
+  senseMap:    SM.State,
+  senseObject: SO.State,
+  selection:   SL.State,
+  editor:      OE.State,
+  input:        I.State,
+  viewport:     V.State,
+  stage:       SG.State,
+  importer:    IP.State,
+};
+
+export type Dispatch = ReduxDispatch<State>;
 
 /**
  * This function applies the dispatch function to every functions in the given object without
